[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 74.2 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,6 +1,6 @@
 class Solution {
-    public int solution(double num1, double num2) {
-        double answer = num1 / num2;
+    public int solution(int num1, int num2) {
+        double answer = (double) num1 / num2;
         return (int) (answer * 1000);
     }
 }
